Breaking Joi validation out into its own module
